public: update basehandle

+++ public/basehandle.h
@@ -63,7 +63,7 @@ public:
 protected:
 	// The low NUM_SERIAL_BITS hold the index. If this value is less than MAX_EDICTS, then the entity is networkable.
 	// The high NUM_SERIAL_NUM_BITS bits are the serial number.
-	unsigned long	m_Index;
+	uint32	m_Index;
 };
